Configured roles: now the project owner can share with other projects

+++ src/main/java/com/zhsaidk/service/CatalogService.java
@@ -1,40 +1,62 @@
 package com.zhsaidk.service;
 
 import com.zhsaidk.database.entity.Catalog;
+import com.zhsaidk.database.entity.PermissionRole;
 import com.zhsaidk.database.entity.Project;
+import com.zhsaidk.database.entity.ProjectPermission;
 import com.zhsaidk.database.repo.CatalogRepository;
 import com.zhsaidk.database.repo.ProjectRepository;
+import com.zhsaidk.database.repo.UserRepository;
 import com.zhsaidk.dto.BuildCreateCatalogDto;
 import com.zhsaidk.dto.BuildReadCatalogDto;
 import com.zhsaidk.util.SlugUtil;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.web.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Objects;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CatalogService {
     private final CatalogRepository catalogRepository;
     private final ProjectRepository projectRepository;
+    private final PermissionService permissionService;
+    private final UserRepository userRepository;
 
-    public PagedModel<Catalog> findAll(PageRequest pageRequest) {
-        return new PagedModel<>(catalogRepository.findAll(pageRequest));
+    public PagedModel<Catalog> findAll(PageRequest pageRequest, Authentication authentication) {
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        return new PagedModel<>(catalogRepository.findAllByUserIdAndAnyRole(pageRequest, userId));
     }
 
-    public Page<Catalog> findAllCatalogs(PageRequest pageRequest){
-        return catalogRepository.findAll(pageRequest);
+    public Page<Catalog> findAllCatalogs(PageRequest pageRequest, Authentication authentication){
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        return catalogRepository.findAllByUserIdAndAnyRole(pageRequest, userId);
     }
 
-    public ResponseEntity<?> findById(String projectSlug, String catalogSlug) {
+    public ResponseEntity<?> findById(String projectSlug, String catalogSlug, Authentication authentication) {
+
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        if (!permissionService.hasAnyPermission(projectSlug, userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         if(!catalogRepository.existsBySlugAndProjectSlug(catalogSlug, projectSlug)){
             return ResponseEntity.badRequest().build();
         }
@@ -44,7 +66,15 @@ public class CatalogService {
         return ResponseEntity.ok(catalog);
     }
 
-    public ResponseEntity<?> build(BuildCreateCatalogDto dto, String projectSlug) {
+    public ResponseEntity<?> build(BuildCreateCatalogDto dto, String projectSlug, Authentication authentication) {
+
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        if (!permissionService.hasPermission(projectSlug, userId, List.of(PermissionRole.OWNER, PermissionRole.WRITER))){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         Project project = projectRepository.findProjectBySlug(projectSlug)
                 .orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "Project not found"));
 
@@ -59,7 +89,16 @@ public class CatalogService {
         return ResponseEntity.status(HttpStatus.CREATED).body(savedCatalog);
     }
 
-    public ResponseEntity<?> update(String projectSlug, String catalogSlug, BuildReadCatalogDto dto) {
+    @Transactional
+    public ResponseEntity<?> update(String projectSlug, String catalogSlug, BuildReadCatalogDto dto, Authentication authentication) {
+
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        if (!permissionService.hasPermission(projectSlug, userId, List.of(PermissionRole.OWNER, PermissionRole.WRITER))){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only owner or writer can change");
+        }
+
         if(!catalogRepository.existsBySlugAndProjectSlug(catalogSlug, projectSlug)){
             return ResponseEntity.badRequest().build();
         }
@@ -81,7 +120,15 @@ public class CatalogService {
     }
 
     @Transactional
-    public Boolean remove(String catalogSlug) {
+    public Boolean remove(String projectSlug, String catalogSlug, Authentication authentication) {
+
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        if (!permissionService.hasPermission(projectSlug, userId, List.of(PermissionRole.OWNER, PermissionRole.WRITER))){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only project owner or writer can delete");
+        }
+
         if (catalogRepository.existsBySlug(catalogSlug)) {
             catalogRepository.deleteBySlug(catalogSlug);
             return true;
@@ -89,12 +136,19 @@ public class CatalogService {
         return false;
     }
 
-    public Catalog findCatalogBySlug(String slug) {
-        return catalogRepository.findCatalogBySlug(slug)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Catalog not found"));
+    public List<Catalog> findAllCatalogsByProjectSlug(String projectSlug, Authentication authentication){
+        Integer userId = userRepository.findUserIdByUsername(authentication.getName())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+
+        if (!permissionService.hasAnyPermission(projectSlug, userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        return catalogRepository.findAllCatalogsByProjectSlug(projectSlug);
     }
 
-    public List<Catalog> findAllCatalogsByProjectSlug(String projectSlug){
-        return catalogRepository.findAllCatalogsByProjectSlug(projectSlug);
+    public Catalog findCatalogByCatalogSlug(String catalogSlug){
+        return catalogRepository.findCatalogBySlug(catalogSlug)
+                .orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "catalog not found"));
     }
 }
